keystroke/client/requirements.txt
websocket-client
tqdm
numpy
